requirements.txt
# Faceless Shorts Automator V2 – Python dependencies
# Install: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Script generation (Gemini)
google-generativeai>=0.3.0

# Voice generation (ElevenLabs + fallback)
elevenlabs>=1.0.0
gTTS>=2.4.0

# Video generation (Runway ML)
runwayml>=1.0.0
httpx>=0.27.0

# Image processing
Pillow>=10.0.0

# YouTube upload
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
